+++ src/skydb.ts
@@ -1,8 +1,8 @@
 import { pki } from "node-forge";
 import { SkynetClient } from "./client";
-import { PublicKey, SecretKey } from "./crypto";
 import { RegistryEntry, SignedRegistryEntry } from "./registry";
-import { parseSkylink, trimUriPrefix, uriSkynetPrefix } from "./utils";
+import { parseSkylink, trimUriPrefix, uriSkynetPrefix, toHexString } from "./utils";
+import { Buffer } from "buffer";
 
 /**
  * Gets the JSON object corresponding to the publicKey and dataKey.
@@ -13,7 +13,7 @@ import { parseSkylink, trimUriPrefix, uriSkynetPrefix } from "./utils";
  */
 export async function getJSON(
   this: SkynetClient,
-  publicKey: PublicKey,
+  publicKey: string,
   dataKey: string,
   customOptions = {}
 ): Promise<{ data: Record<string, unknown>; revision: number } | null> {
@@ -41,9 +41,12 @@ export async function getJSON(
   return { data: response.data, revision: entry.entry.revision };
 }
 
+/**
+ * Sets a JSON object at the registry entry corresponding to the publicKey and dataKey.
+ */
 export async function setJSON(
   this: SkynetClient,
-  privateKey: SecretKey,
+  privateKey: string,
   dataKey: string,
   json: Record<string, unknown>,
   revision?: number,
@@ -54,18 +57,19 @@ export async function setJSON(
     ...customOptions,
   };
 
+  const privateKeyBuffer = Buffer.from(privateKey, "hex");
+
   // Upload the data to acquire its skylink
   // TODO: Replace with upload request method.
   const file = new File([JSON.stringify(json)], dataKey, { type: "application/json" });
   const { skylink } = await this.uploadFileRequest(file, opts);
 
-  if (!revision) {
+  if (revision === undefined) {
     // fetch the current value to find out the revision.
     let entry: SignedRegistryEntry;
     try {
-      const publicKey = pki.ed25519.publicKeyFromPrivateKey({ privateKey });
-      entry = await this.registry.getEntry(publicKey, dataKey, opts);
-
+      const publicKey = pki.ed25519.publicKeyFromPrivateKey({ privateKey: privateKeyBuffer });
+      entry = await this.registry.getEntry(toHexString(publicKey), dataKey, opts);
       revision = entry.entry.revision + 1;
     } catch (err) {
       revision = 0;
@@ -80,5 +84,5 @@ export async function setJSON(
   };
 
   // update the registry
-  await this.registry.setEntry(privateKey, dataKey, entry);
+  await this.registry.setEntry(privateKey, entry);
 }